// client/src/components/user/__generated__/UserList_query.graphql.ts
export type ClogUserRole = "EDITOR" | "READER" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type UserList_query$data = {
  readonly clogUsers: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly isAdmin: boolean;
        readonly isEnabled: boolean;
        readonly isViewer: boolean;
        readonly role: ClogUserRole;
        readonly username: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "UserList_query";
};
export type UserList_query$key = {
  readonly " $data"?: UserList_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"UserList_query">;
};